requirements.txt
beautifulsoup4
feedgen

python-dateutil
pytz

